% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_taxa.R
\name{allTaxa}
\alias{allTaxa}
\title{All taxa}
\usage{
allTaxa(taxa = NULL, file = NULL, return = TRUE, overwrite = FALSE,
  memoised = FALSE)
}
\arguments{
\item{taxa}{Limits the result to entities with given identifiers.

A list of identifiers, separated by commas (e.g: human, mouse, fly).
 Identifiers can be the any of the following:

\itemize{
  \item taxon ID
  \item scientific name
  \item common name
  \item abbreviation
}}

\item{file}{Character. File path. If provided, response will be saved to file}

\item{return}{Logical. If the response should be returned. Set to false when
you only want to save a file}

\item{overwrite}{Logical. If TRUE, existing files will be overwritten. If FALSE
a warning will be thrown and no action is taken.}

\item{memoised}{Logical. If TRUE a memoised version of the function will be
used which is faster for repeated requests. Use \code{\link{forgetGemmaMemoised}}
to clear memory.}
}
\value{
List of lists containing experiment object.
}
\description{
Lists all available taxa.
}
\examples{
allTaxa()
allTaxa(c('human','mouse'))
}
