% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annotationGeneMatch}
\alias{annotationGeneMatch}
\title{Match probesets to gene in a platform annotation}
\usage{
annotationGeneMatch(probesets, annotation, removeNAs = FALSE)
}
\arguments{
\item{probesets}{A character vector of probeset names}

\item{annotation}{Output of \code{\link{getAnnotation}} or a file path to an annotation}

\item{removeNAs}{If no match is found, NAs will be return. This removes NAs
information written by getAnnotation}
}
\value{
A character vector of gene names
}
\description{
Match probesets to gene in a platform annotation
}
