% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_other.R
\name{annotationInfo}
\alias{annotationInfo}
\title{annotationInfo}
\usage{
annotationInfo(annotation, request = NULL, file = NULL, return = TRUE,
  overwrite = FALSE, memoised = FALSE)
}
\arguments{
\item{annotation}{Character. Either plain text, ontology term ID or an ontology term URI.}

\item{request}{Character. If null retreeives the annotation object. Otherwise
\itemize{
    \item \code{datasets}: Does a search for datasets containing annotations matching the given string
}}

\item{file}{Character. File path. If provided, response will be saved to file}

\item{return}{Logical. If the response should be returned. Set to false when
you only want to save a file}

\item{overwrite}{Logical. If TRUE, existing files will be overwritten. If FALSE
a warning will be thrown and no action is taken.}

\item{memoised}{Logical. If TRUE a memoised version of the function will be
used which is faster for repeated requests. Use \code{\link{forgetGemmaMemoised}}
to clear memory.}
}
\value{
A list
}
\description{
annotationInfo
}
\examples{
annotationInfo('http://purl.obolibrary.org/obo/OBI_0000105')
annotationInfo('transplantation')
annotationInfo('http://purl.obolibrary.org/obo/OBI_0000105', request = 'datasets')
}
