% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annotationProbesetMatch}
\alias{annotationProbesetMatch}
\title{Match probesets to gene in a platform annotation}
\usage{
annotationProbesetMatch(genes, annotation, removeNAs = FALSE)
}
\arguments{
\item{genes}{A character vector of gene names}

\item{annotation}{Output of \code{\link{getMicroannot}} or a file path to an annotation
information written by getMicroannot}

\item{removeNAs}{If no match is found, NAs will be return. This removes NAs}
}
\value{
A character vector of probeset names
}
\description{
Match probesets to gene in a platform annotation
}
