% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highLevelFunctions.R
\name{compileMetadata}
\alias{compileMetadata}
\title{compileMetadata}
\usage{
compileMetadata(dataset, collapseBioMaterials = TRUE,
  outputType = c("data.frame", "list"), memoised = TRUE)
}
\arguments{
\item{dataset}{Character. Can either be the dataset ID or its short name 
(e.g. GSE1234).}

\item{collapseBioMaterials}{If TRUE collapses samples into biomaterials
Some datasets represent}

\item{outputType}{data.frame or list. A data.frame will return a single data frame
including experiment and sample data. List output will return a list of length 2 
where experiment and sample data are separated.}

\item{memoised}{Should memoisation be used for base API calls. Useful if calling
a large number of datasets in succession, defaults to TRUE. 
Use \code{\link{forgetGemmaMemoised}} to clear memory.}
}
\value{
A data.frame.
}
\description{
Compiles metadata for a single dataset that includes sample and sample
metadata. Note that this function uses memoised functions to speed up loops.
If something changes in gemma as you work and you want to update, restart 
your session or use \code{\link{forgetGemmaMemoised}}
}
\examples{
compileMetadata('GSE11322')
compileMetadata(1017)
}
