% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{downloadDataset}
\alias{downloadDataset}
\title{Download Expression file for a dataset}
\usage{
downloadDataset(dataset, file = NULL, filter = FALSE, return = TRUE)
}
\arguments{
\item{dataset}{Character. Can either be the ExpressionExperiment ID or its 
short name (e.g. GSE1234). Retrieval by ID is more efficient. Only datasets
that user has access to will be available}

\item{file}{Character. File path. If provided expression data will be saved.}

\item{filter}{Logical. If true filters the expression data so that they do
not contain samples under a minimum threshold.}

\item{return}{Logical. If true returns the expression data.}
}
\value{
A data.frame or NULL depending on return value.
}
\description{
Given an experiment, returns and/or saves the expression data.
}
