% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationDirectory.R
\name{endpointFunctions}
\alias{endpointFunctions}
\title{Endpoint functions}
\description{
Basic API calls. These functions return mostly unaltered data from a given API endpoint.
}
\details{
\link{allDatasets}: List all datasets

\link{datasetInfo}: Get information about a specific dataset

\link{allPlatforms}: List all platforms

\link{platformInfo}: Get information about a specific platform

\link{geneInfo}: Get information about genes

\link{allTaxa}: List all taxa

\link{taxonInfo}: Get information about a specific taxon

\link{annotationInfo}: Query for annotations
}
