% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{geneInfo}
\alias{geneInfo}
\title{geneInfo}
\usage{
geneInfo(gene, request = NULL, ..., file = NULL, return = TRUE)
}
\arguments{
\item{gene}{Can either be the NCBI ID (1859), Ensembl ID (ENSG00000157540) or official symbol (DYRK1A) of the gene. 

NCBI ID is the most efficient (and guaranteed to be unique) identifier.

Official symbol represents a gene homologue for a random taxon, unless used in a specific taxon (see the Taxa Endpoints).}

\item{request}{Character. If NULL retrieves all genes matching the identifier. Otherwise
\itemize{
    \item \code{evidence}: Retrieves gene evidence for the given gene
    \item \code{locations}: Retrieves the physical location of the given gene
    \item \code{probes}: Retrieves the probes (composite sequences) with this gene. Parameters:
        \itemize{
            \item \code{offset}: Optional, defaults to 0. Skips the 
            specified amount of objects when retrieving them from the
            database.
            \item \code{limit}: Optional, defaults to 20. Limits the result 
            to specified amount of objects. Use 0 for no limit.
        }
    \item \code{goTerms}: Retrieves the GO terms of the given gene
    \item \code{coexpression}: Retrieves the coexpression of two given genes. Parameters:
        \itemize{
            \item \code{with}: Required. Can either be the NCBI ID (1859), Ensembl ID (ENSG00000157540) or official symbol (DYRK1A) of the gene.
            
            NCBI ID is the most efficient (and guaranteed to be unique) identifier.
            
            Official symbol represents a gene homologue for a random taxon, unless used in a specific taxon (see the Taxa Endpoints).
            \item \code{limit}: Optional, defaults to 100. 
            
            Limits the result to specified amount of objects. Use 0 for no limit.
            \item \code{stringency}: Optional, defaults to 1. Sets the stringency of coexpression search.
        }
}}

\item{...}{Use if the specified request has additional parameters.}

\item{file}{Character. File path. If provided, response will be saved to file}

\item{return}{Logical. If the response should be returned. Set to false when
you only want to save a file}
}
\value{
A list
}
\description{
geneInfo
}
\examples{
geneInfo('1859') # single match it is a unique NCBI id
geneInfo('DYRK1A') # this returns all genes named DYRK1A from all species
}
