% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{getAnnotation}
\alias{getAnnotation}
\title{Get Gemma annotations}
\usage{
getAnnotation(
  platform,
  annotType = c("bioProcess", "noParents", "allParents"),
  file = NULL,
  return = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{platform}{Can either be the platform ID or its short name (e.g: GPL1355). 
Retrieval by ID is more efficient. Only platforms that user has access to will be available.}

\item{annotType}{Which go terms should the input include}

\item{file}{Character. File path. If provided, response will be saved to file}

\item{return}{Logical. If the response should be returned. Set to false when
you only want to save a file}

\item{overwrite}{Logical. If TRUE, existing files will be overwritten. If FALSE
a warning will be thrown and no action is taken.}
}
\value{
A data frame including annotation information
}
\description{
Gets Gemma's platform annotation files that can be accessed from https://gemma.msl.ubc.ca/annots/
}
