% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higherLevelFunctions.R
\name{getDatasets}
\alias{getDatasets}
\title{Get list of datasets}
\usage{
getDatasets(taxa = NULL, platforms = NULL)
}
\arguments{
\item{taxa}{Character vector. Can either be Taxon ID or one of its string
identifiers: scientific name, common name, abbreviation. It is recommended
to use ID for efficiency.}

\item{platforms}{Character vector. Can either be the ArrayDesign ID or its
short name (e.g. "Generic_yeast" or "GPL1355" ). Retrieval by ID is more
efficient. Only platforms that user has access to will be available.}
}
\value{
List of lists containing experiment data.
}
\description{
Outputs lists of datasets optionally filtered based on taxa and platform.
}
