% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{getTaxa}
\alias{getTaxa}
\title{List all taxa or return specific taxa}
\usage{
getTaxa(taxon = NULL)
}
\arguments{
\item{taxon}{Character. can either be Taxon ID or one of its string
identifiers: scientific name, common name, abbreviation. It is recommended
to use ID for efficiency.}
}
\description{
List all taxa or return specific taxa
}
