% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationDirectory.R
\name{highLevelFunctions}
\alias{highLevelFunctions}
\title{High Level Functions}
\description{
High level functions that create useful outputs from basic API calls or from their
outputs
}
\details{
\code{\link{compileMetadata}}

\code{\link{expressionSubset}}

\code{\link{readExpression}}
}
