% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonArguments.R
\name{memoised}
\alias{memoised}
\title{memoised}
\arguments{
\item{memoised}{Logical. If TRUE a memoised version of the function will be
used which is faster for repeated requests. Use \link{\code{forgetGemmaMemoised}}
to clear memory.}
}
\description{
memoised
}
\keyword{internal}
