% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_platform.R
\name{platformInfo}
\alias{platformInfo}
\title{platformInfo}
\usage{
platformInfo(
  platform,
  request = NULL,
  ...,
  file = NULL,
  return = TRUE,
  overwrite = FALSE,
  memoised = FALSE
)
}
\arguments{
\item{platform}{Can either be the platform ID or its short name (e.g: GPL1355). 
Retrieval by ID is more efficient. Only platforms that user has access to will be available.

If a vector of length>1 is provided return all matching platform
objects similar to \code{\link{allPlatforms}} but without access to additional 
parameters. \code{request} parameter cannot be specified for vector inputs}

\item{request}{Character. If NULL retrieves the platform object. Otherwise
\itemize{
    \item \code{datasets}: Retrieves experiments in the given platform. Parameters:
        \itemize{
            \item \code{offset}: Optional, defaults to 0. Skips the 
             specified amount of objects when retrieving them from the
              database.
            \item \code{limit}: Optional, defaults to 20. Limits the result 
             to specified amount of objects. Use 0 for no limit.
        }
    \item \code{annotations}: Retrieves the annotation file for the given platform. If you set a file path, the downloaded file will be a .gz file.
    \item \code{elements}: Retrieves the composite sequences (elements/probes) for the given platform. Parameters:
        \itemize{
            \item \code{offset}: As in datasets
            \item \code{limit}: As in datasets
        }
    \item \code{specificElement} Retrieves a specific composite sequence (element) for the given platform. Paramaters: 
        \itemize{
            \item \code{probe}: Required. Can either be the probe name or ID.
        }
    \item \code{genes} Retrieves the genes on the given platform element. Parameters:
        \itemize{
            \item \code{probe}: Required. Can either be the probe name or ID.
            \item \code{offset}: As in datasets
            \item \code{limit}: As in datasets
        }
}}

\item{...}{Use if the specified request has additional parameters.}

\item{file}{Character. File path. If provided, response will be saved to file}

\item{return}{Logical. If the response should be returned. Set to false when
you only want to save a file}

\item{overwrite}{Logical. If TRUE, existing files will be overwritten. If FALSE
a warning will be thrown and no action is taken.}

\item{memoised}{Logical. If TRUE a memoised version of the function will be
used which is faster for repeated requests. Use \code{\link{forgetGemmaMemoised}}
to clear memory.}
}
\value{
A data.frame or a list depending on the request
}
\description{
Retrieves information about a single platform. Combines several API calls.
}
\examples{
\donttest{
platformInfo('GPL1355')
platformInfo('GPL1355',request = 'datasets',limit = 10)
}
}
