% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonArguments.R
\name{queryLimit}
\alias{queryLimit}
\title{queryLimit}
\usage{
queryLimit(offset = 0, limit = 20)
}
\arguments{
\item{offset}{Integer. Optional parameter (defaults to 0) skips the specified amount of datasets when retrieving them from the database.}

\item{limit}{Integer. Optional parameter (defaults to 20) limits the result to specified amount of datasets. Use 0 for no limit.}
}
\description{
queryLimit
}
\keyword{internal}
