% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highLevelFunctions.R
\name{readExpression}
\alias{readExpression}
\title{Read expression data}
\usage{
readExpression(expFile, IdColnames = FALSE)
}
\arguments{
\item{IdColnames}{Logical. should column names be turned into IDs}

\item{gzFile}{Path to expression file downloaded from gemma, possibly using datasetInfo
function (request = 'data'). Should be in its original gzipped form}
}
\value{
A data frame
}
\description{
Read expression data
}
