% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{setGemmaUser}
\alias{setGemmaUser}
\title{Gemma authentication}
\usage{
setGemmaUser(username = NULL, password = NULL)
}
\arguments{
\item{username}{Gemma username}

\item{password}{Gemma password}
}
\description{
Allows the user to access information that requires logging in to Gemma. To log out, run
\code{setGemmaUser} without specifying the username or password
}
\examples{
setGemmaUser('username','password') # login
setGemmaUser() # log out
}
