% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonArguments.R
\name{sortArg}
\alias{sortArg}
\title{Sort argument}
\usage{
sortArg(sort)
}
\arguments{
\item{sort}{Character. Optional parameter (defaults to +id) sets the ordering property
and direction. Format is [+,-][property name]. E.g. "-accession" will
translate to descending ordering by the Accession property. Note that this
does not guarantee the order of the returned entities. Nested properties
are also supported (recursively). 

E.g:
+curationDetails.lastTroubledEvent.date}
}
\description{
Sort argument
}
\keyword{internal}
