% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highLevelFunctions.R
\name{splitExpression}
\alias{splitExpression}
\title{Split expression data into gene metadata expression data}
\usage{
splitExpression(expression)
}
\arguments{
\item{expression}{Expression data outputted by readExpression/readDataFile}
}
\description{
Split expression data into gene metadata expression data
}
