% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{allTaxa}
\alias{allTaxa}
\title{All taxa}
\usage{
allTaxa(file = NULL, return = TRUE)
}
\arguments{
\item{file}{Character. File path. If provided, response will be saved to file}

\item{return}{Logical. If the response should be returned. Set to false when
you only want to save a file}
}
\value{
List of lists containing experiment object.
}
\description{
Lists all available taxa.
}
\examples{
allTaxa()
}
