% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{datasetInfo}
\alias{datasetInfo}
\title{datasetInfo}
\usage{
datasetInfo(dataset, request = NULL, ..., file = NULL)
}
\arguments{
\item{dataset}{Character. Can either be the dataset ID or its short name 
(e.g. GSE1234).}

\item{request}{Character. If NULL retrieves the dataset object. Otherwise
 \itemize{
     \item \code{platforms}: Retrieves platforms for the given dataset
     \item \code{samples}: Retrieves samples for the given dataset
     \item \code{annotations}: Retrieves the annotations for the given dataset
     \item \code{design}: Retrieves the design for the given dataset
     \item \code{data}: Retrieves the data for the given dataset. Parameters:
         \itemize{
             \item \code{filter}: Optional, defaults to false. If true, call
              returns filtered expression data.
         }
     \item \code{differential}: Retrieves the differential analysis results
      for the given dataset. Parameters:
         \itemize{
             \item \code{qValueThreshold}: Required. Q-value threshold.
             \item \code{offset}: Optional, defaults to 0. Skips the specified amount of objects when retrieving them from the database.
             \item \code{limit}: Optional, defaults to 20. Limits the result to specified amount of objects. Use 0 for no limit.
         }
}}

\item{...}{Use if the specified request has additional parameters.}

\item{file}{Character. File path. If provided, response will be saved to file}
}
\description{
Retrieves information about a single dataset based on the given dataset 
identifier. Combines several API calls
}
\examples{
datasetInfo('GSE81454')
}
