% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationDirectory.R
\name{endpointFunctions}
\alias{endpointFunctions}
\title{Endpoint functions}
\description{
Basic API calls. These functions return mostly unaltered data from a given API endpoint.
}
\details{
\code{\link{setGemmaUser}}: Allows authentication

\code{\link{allDatasets}}: List all datasets

\code{\link{datasetInfo}}: Get information about a specific dataset

\code{\link{allPlatforms}}: List all platforms

\code{\link{platformInfo}}: Get information about a specific platform

\code{\link{geneInfo}}: Get information about genes

\code{\link{allTaxa}}: List all taxa

\code{\link{taxonInfo}}: Get information about a specific taxon

\code{\link{annotationInfo}}: Query for annotations

\code{\link{setGemmaUser}}: Sets gemma username and password for authentication
}
