% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonArguments.R
\name{fileReturn}
\alias{fileReturn}
\title{file and return}
\arguments{
\item{file}{Character. File path. If provided, response will be saved to file}

\item{return}{Logical. If the response should be returned. Set to false when
you only want to save a file}
}
\description{
file and return
}
\keyword{internal}
